requirements.txt
deepgram-sdk==3.11.0
Flask==3.1.0
huggingface-hub==0.30.2
gunicorn==22.0.0
numpy==2.2.4
openai==1.74.0
pandas==2.2.3
pillow==11.2.1
PyMuPDF==1.25.5
pytesseract==0.3.13
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sympy==1.13.1
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
